updated price to float in add product

+++ client/src/utils/mutations.js
@@ -25,7 +25,7 @@ export const ADD_USER = gql`
 `;
 
 export const ADD_PRODUCT = gql`
-  mutation AddProduct($name: String!, $description: String!, $price: Int!, $image: String!) {
+  mutation AddProduct($name: String!, $description: String!, $price: Float!, $image: String!) {
   addProduct(name: $name, description: $description, price: $price, image: $image) {
     _id
     name
